backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ML/Computer Vision
torch==2.5.0
torchvision==0.20.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# Data validation
pydantic==2.9.0
pydantic-settings==2.5.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
